// src/api/models/ticketListItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Allta Partners
 * Allta Partners 서버 문서입니다.
 * OpenAPI spec version: 1.0
 */
import type { TicketUserItem } from './ticketUserItem';
import type { TicketStoreItem } from './ticketStoreItem';

export interface TicketListItem {
  /** ID */
  id: string;
  /** 일회권 상태 */
  status: string;
  /** 서비스 종류 */
  serviceType: string;
  /** 차량 번호 */
  carNumber: string;
  /** 생성일 */
  createdAt: string;
  /** 만료일 */
  expiredAt: string;
  /** 회원 정보 */
  user: TicketUserItem;
  /** 매장 정보 */
  store: TicketStoreItem;
}